extract OKF into a dedicated src/okf module
