--- src/agent/frontmatter-validator.ts
@@ -1,211 +1,0 @@
-import { ToolMessage } from "@langchain/core/messages";
-import type { BackendProtocolV2 } from "deepagents";
-import path from "node:path";
-import { parse } from "yaml";
-import { MUTATION_PATH_METADATA_KEY } from "./docs-only-backend.js";
-import type { OpenWikiOutputMode } from "./types.js";
-
-const OKF_STRING_FIELDS = [
-  "type",
-  "title",
-  "description",
-  "resource",
-  "timestamp",
-];
-const OKF_RESERVED_FILES = new Set(["index.md", "log.md"]);
-const WRITE_TOOLS = new Set(["write_file", "edit_file"]);
-
-interface FrontmatterIssue {
-  code: string;
-  line?: number;
-  message: string;
-}
-
-export type FrontmatterValidation =
-  | {
-      valid: true;
-    }
-  | {
-      issues: FrontmatterIssue[];
-      valid: false;
-    };
-
-/** Parses and validates OKF front matter while tolerating producer extensions. */
-export function validateOkfFrontmatter(content: string): FrontmatterValidation {
-  const lines = content.split(/\r?\n/u);
-  if (lines[0] !== "---") {
-    return invalid(
-      "missing_opening_delimiter",
-      "File must begin with `---`.",
-      1,
-    );
-  }
-
-  const closingLine = lines.indexOf("---", 1);
-  if (closingLine === -1) {
-    return invalid(
-      "missing_closing_delimiter",
-      "Opening front matter has no closing `---` delimiter.",
-    );
-  }
-
-  let fields: unknown;
-  try {
-    fields = parse(`\n${lines.slice(1, closingLine).join("\n")}`, {
-      maxAliasCount: 100,
-      schema: "core",
-      uniqueKeys: true,
-    }) as unknown;
-  } catch (error) {
-    return invalid("invalid_yaml", errorMessage(error));
-  }
-  if (!isRecord(fields)) {
-    return invalid("invalid_yaml_root", "Front matter must be a YAML mapping.");
-  }
-
-  const issues: FrontmatterIssue[] = [];
-
-  if (!Object.hasOwn(fields, "type")) {
-    issues.push(issue("missing_type", "Required field `type` is missing."));
-  }
-  for (const field of OKF_STRING_FIELDS) {
-    if (
-      Object.hasOwn(fields, field) &&
-      (typeof fields[field] !== "string" || !fields[field].trim())
-    ) {
-      issues.push(
-        issue(
-          `invalid_${field}`,
-          `Field \`${field}\` must be a non-empty string.`,
-        ),
-      );
-    }
-  }
-  if (
-    Object.hasOwn(fields, "tags") &&
-    (!Array.isArray(fields.tags) ||
-      fields.tags.some((tag) => typeof tag !== "string" || !tag.trim()))
-  ) {
-    issues.push(
-      issue(
-        "invalid_tags",
-        "Field `tags` must be a YAML list of non-empty strings.",
-      ),
-    );
-  }
-
-  return issues.length === 0 ? { valid: true } : { issues, valid: false };
-}
-
-/** Appends an actionable warning when a wiki write leaves invalid front matter. */
-export async function addFrontmatterWarning<Result>(
-  result: Result,
-  backend: BackendProtocolV2,
-  outputMode: OpenWikiOutputMode,
-  toolName: string,
-): Promise<Result> {
-  if (!WRITE_TOOLS.has(toolName)) return result;
-
-  const mutation = getToolMessages(result)
-    .map((message) => ({
-      message,
-      path: message.metadata?.[MUTATION_PATH_METADATA_KEY],
-    }))
-    .find(
-      (item): item is { message: ToolMessage; path: string } =>
-        typeof item.path === "string" &&
-        isWikiMarkdownPath(item.path, outputMode),
-    );
-  if (!mutation) return result;
-
-  const validation = await validatePersistedFile(backend, mutation.path);
-  if (validation.valid) return result;
-
-  const warning = formatWarning(mutation.path, validation.issues);
-  mutation.message.content =
-    typeof mutation.message.content === "string"
-      ? `${mutation.message.content}\n\n${warning}`
-      : [...mutation.message.content, { text: warning, type: "text" }];
-  return result;
-}
-
-/** Reads a persisted Markdown file and validates its final front matter. */
-async function validatePersistedFile(
-  backend: BackendProtocolV2,
-  filePath: string,
-): Promise<FrontmatterValidation> {
-  const read = await backend.readRaw(filePath);
-  const content = read.data?.content;
-  if (read.error || content === undefined || content instanceof Uint8Array) {
-    return invalid(
-      "file_read_failed",
-      `Could not read the final Markdown text: ${read.error ?? "no text data"}.`,
-    );
-  }
-  return validateOkfFrontmatter(
-    Array.isArray(content) ? content.join("\n") : content,
-  );
-}
-
-/** Extracts tool messages from direct and Command-like tool results. */
-function getToolMessages(result: unknown): ToolMessage[] {
-  if (ToolMessage.isInstance(result)) return [result];
-  if (!isRecord(result)) return [];
-
-  const messages = isRecord(result.update) ? result.update.messages : undefined;
-  return Array.isArray(messages)
-    ? messages.filter((message): message is ToolMessage =>
-        ToolMessage.isInstance(message),
-      )
-    : [];
-}
-
-/** Checks whether a path targets an OKF concept document inside the wiki. */
-function isWikiMarkdownPath(
-  filePath: string,
-  outputMode: OpenWikiOutputMode,
-): boolean {
-  const normalized = path.posix.normalize(
-    `/${filePath.trim().replaceAll("\\", "/").replace(/^\/+/, "")}`,
-  );
-  return (
-    path.posix.extname(normalized).toLowerCase() === ".md" &&
-    !OKF_RESERVED_FILES.has(path.posix.basename(normalized).toLowerCase()) &&
-    (outputMode === "local-wiki" || normalized.startsWith("/openwiki/"))
-  );
-}
-
-/** Formats validation issues as an instruction for the agent to correct the file. */
-function formatWarning(path: string, issues: FrontmatterIssue[]): string {
-  const details = issues
-    .map(
-      ({ code, line, message }) =>
-        `- [${code}]${line ? ` line ${line}:` : ""} ${message}`,
-    )
-    .join("\n");
-  return `WARNING: YAML front matter was NOT formatted properly in \`${path}\`.\n${details}\nYou MUST correct this file's YAML front matter before continuing.`;
-}
-
-/** Creates a failed validation result containing one issue. */
-function invalid(
-  code: string,
-  message: string,
-  line?: number,
-): FrontmatterValidation {
-  return { issues: [issue(code, message, line)], valid: false };
-}
-
-/** Creates a structured front-matter validation issue. */
-function issue(code: string, message: string, line?: number): FrontmatterIssue {
-  return { code, ...(line ? { line } : {}), message };
-}
-
-/** Narrows an unknown value to a non-array object record. */
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return value !== null && typeof value === "object" && !Array.isArray(value);
-}
-
-/** Converts an unknown thrown value into a readable message. */
-function errorMessage(error: unknown): string {
-  return error instanceof Error ? error.message : String(error);
-}
